+++ client/src/pages/sets/[id]/flashcards.tsx
@@ -1,8 +1,8 @@
 import { GetServerSideProps } from 'next';
 import { QueryClient, dehydrate } from 'react-query';
 import { setApi } from '@src/apis';
+import { consts } from '@src/lib';
 import Flashcards from '@src/modules/set/Flashcards';
-import consts from '@src/lib/consts';
 
 export const getServerSideProps: GetServerSideProps = async (context) => {
   const queryId = typeof context.query.id === 'string' ? context.query.id : '';
